requirements.txt
Flask==3.0.3
gunicorn==23.0.0
line-bot-sdk==3.13.0
openai==1.44.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.0
